Remove DeleteCommand abstraction

+++ src/main/java/seedu/address/logic/parser/DeleteCommandParser.java
@@ -6,17 +6,21 @@ import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.address.logic.commands.DeleteAttributeCommand;
-import seedu.address.logic.commands.DeleteCommand;
-import seedu.address.logic.commands.DeleteIntervieweeCommand;
-import seedu.address.logic.commands.DeleteMetricCommand;
-import seedu.address.logic.commands.DeleteQuestionCommand;
+import seedu.address.logic.commands.*;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
  * Parses input arguments and creates a new DeleteCommand object
  */
-public class DeleteCommandParser implements Parser<DeleteCommand> {
+public class DeleteCommandParser implements Parser<Command> {
+
+    private static final String INPUT_WORD = "delete";
+
+    private static final String INPUT_FORMAT = INPUT_WORD + ": Deletes the properties of an interview session "
+            + "Includes: "
+            + "deleting an Interviewee object"
+            + "deleting an Attribute object"
+            + "deleting a Question object";
 
     private static final Pattern BASIC_DELETE_COMMAND_FORMAT =
             Pattern.compile("(?<deleteCommandWord>\\S+) (?<deleteArguments>.+)");
@@ -28,23 +32,29 @@ public class DeleteCommandParser implements Parser<DeleteCommand> {
      * @param arguments the arguments to be parsed
      * @throws ParseException if the user input does not conform the expected format
      */
-    public DeleteCommand parse(String arguments) throws ParseException {
+    public Command parse(String arguments) throws ParseException {
         Matcher matcher = BASIC_DELETE_COMMAND_FORMAT.matcher(arguments.trim());
 
         if (!matcher.matches()) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE));
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, INPUT_FORMAT));
         }
         final String deleteCommandWord = matcher.group("deleteCommandWord");
         final String deleteArguments = matcher.group("deleteArguments");
 
         switch (deleteCommandWord) {
         case DeleteAttributeCommand.COMMAND_WORD:
+            if (ParserUtil.isEmptyArgument(arguments)) {
+                throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteAttributeCommand.MESSAGE_USAGE));
+            }
             return new DeleteAttributeCommand(deleteArguments.trim());
 
         case DeleteIntervieweeCommand.COMMAND_WORD:
             return new DeleteIntervieweeCommand(deleteArguments.trim());
 
         case DeleteQuestionCommand.COMMAND_WORD:
+            if (ParserUtil.isEmptyArgument(arguments)) {
+                throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteQuestionCommand.MESSAGE_USAGE));
+            }
             return new DeleteQuestionCommand(deleteArguments.trim());
 
         case DeleteMetricCommand.COMMAND_WORD:
